Send SNS notifications from a new service
